+++ docs/changelog.rst
@@ -4,6 +4,9 @@ Changelog
 `Unreleased`_ - TBD
 -------------------
 
+`3.11.2`_ - 2021-11-30
+----------------------
+
 **Changed**
 
 - Use ``name`` & ``data_generation_method`` parameters to subtest context instead of ``path`` & ``method``.
@@ -2022,7 +2025,8 @@ Deprecated
 
 - Initial public release
 
-.. _Unreleased: https://github.com/schemathesis/schemathesis/compare/v3.11.1...HEAD
+.. _Unreleased: https://github.com/schemathesis/schemathesis/compare/v3.11.2...HEAD
+.. _3.11.2: https://github.com/schemathesis/schemathesis/compare/v3.11.1...v3.11.2
 .. _3.11.1: https://github.com/schemathesis/schemathesis/compare/v3.11.0...v3.11.1
 .. _3.11.0: https://github.com/schemathesis/schemathesis/compare/v3.10.1...v3.11.0
 .. _3.10.1: https://github.com/schemathesis/schemathesis/compare/v3.10.0...v3.10.1